features as per plan
